requirements.txt
blinker==1.4
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-instagram==1.3.2
simplejson==3.8.0
six==1.9.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
wheel==0.24.0
